попытка хранить размеры изображений в названии фалов уперлась в отсутстиве в базе имени для medium размера

+++ src/main/java/ru/org/linux/gallery/Screenshot.java
@@ -16,24 +16,31 @@
 package ru.org.linux.gallery;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.validation.Errors;
 import ru.org.linux.util.BadImageException;
 import ru.org.linux.util.images.ImageCheck;
 import ru.org.linux.util.UtilException;
+import ru.org.linux.util.images.ImageInfo;
 import ru.org.linux.util.images.ImageUtil;
 
 import java.io.File;
 import java.io.IOException;
 
 public class Screenshot {
+  private static final Log logger = LogFactory.getLog(Screenshot.class);
   public static final int MAX_SCREENSHOT_FILESIZE = 1500000;
   public static final int MIN_SCREENSHOT_SIZE = 400;
   public static final int MAX_SCREENSHOT_SIZE = 3000;
 
-  private final File mainFile;
-  private final File mediumFile;
-  private final File iconFile;
-  private final String extension;
+  private final String name;
+  private final String ext;
+  private final String path;
+
+  private String fullName;
+  private String mediumName;
+  private String iconName;
 
   private static final int ICON_WIDTH = 200;
   private static final int MEDIUM_WIDTH = 500;
@@ -88,28 +95,23 @@ public class Screenshot {
   }
 
   private Screenshot(String name, String path, String extension) {
-    String mainname = name + '.' + extension;
-    String iconname = name + "-icon.jpg";
-    String medname = name + "-med.jpg";
-
-    mainFile = new File(path, mainname);
-    iconFile = new File(path, iconname);
-    mediumFile = new File(path, medname);
-
-    this.extension = extension;
+    this.name = name;
+    this.path = path;
+    this.ext = extension;
   }
 
-  public Screenshot moveTo(String dir, String name) throws IOException {
-    Screenshot dest = new Screenshot(name, dir, extension);
-
-    FileUtils.moveFile(mainFile, dest.mainFile);
-    FileUtils.moveFile(iconFile, dest.iconFile);
-    FileUtils.moveFile(mediumFile, dest.mediumFile);
-
+  public Screenshot moveTo(String dir, String name) throws IOException, UtilException, BadImageException {
+    Screenshot dest = new Screenshot(name, dir, ext);
+    dest.doResize(new File(fullName));
+    new File(mediumName).delete();
+    new File(iconName).delete();
     return dest;
   }
 
   private void doResize(File uploadedFile) throws IOException, UtilException, BadImageException {
+    ImageCheck check = ImageUtil.imageCheck(uploadedFile);
+    File mainFile = new File(path, name + "-w" + check.getWidth() + "-h" + check.getHeight() + '.' + check.getExtension());
+    fullName = mainFile.getAbsolutePath();
     if (mainFile.exists()) {
       mainFile.delete();
     }
@@ -118,21 +120,29 @@ public class Screenshot {
 
     boolean error = true;
 
+    File mediumFile=null;
+    File iconFile=null;
+
     try {
-      ImageUtil.resizeImage(mainFile.getAbsolutePath(), iconFile.getAbsolutePath(), ICON_WIDTH);
-      ImageUtil.resizeImage(mainFile.getAbsolutePath(), mediumFile.getAbsolutePath(), MEDIUM_WIDTH);
+      iconName = ImageUtil.resizeImage(mainFile.getAbsolutePath(), name, path, "-icon", ICON_WIDTH);
+      mediumName = ImageUtil.resizeImage(mainFile.getAbsolutePath(), name, path, "-med", MEDIUM_WIDTH);
+      iconFile = new File(iconName);
+      mediumFile = new File(mediumName);
       error = false;
+    } catch (Exception e) {
+      logger.debug(e.getMessage());
     } finally {
       if (error) {
+        logger.debug("error move");
         if (mainFile.exists()) {
           mainFile.delete();
         }
 
-        if (iconFile.exists()) {
+        if (iconFile != null && iconFile.exists()) {
           iconFile.delete();
         }
 
-        if (mediumFile.exists()) {
+        if (mediumFile != null && mediumFile.exists()) {
           iconFile.delete();
         }
       }
@@ -140,10 +150,10 @@ public class Screenshot {
   }
 
   public File getMainFile() {
-    return mainFile;
+    return new File(fullName);
   }
 
   public File getIconFile() {
-    return iconFile;
+    return new File(iconName);
   }
 }
